% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasthedata.R
\name{seasthedata}
\alias{seasthedata}
\title{Easily seasonally adjust a tibble or data.frame of long data}
\usage{
seasthedata(original, frequency = NULL, use_original = FALSE, ...)
}
\arguments{
\item{original}{Your data, in original terms, to be seasonally adjusted.}

\item{frequency}{(Optional) The frequency of your data. Accepted options are
"quarter", "month", "day". If omitted, seasthedata will guess the frequency.}

\item{use_original}{(default: FALSE) If the series cannot be seasonally
adjusted, should the returned data have NAs for the relevant series (default)
or the original data.}

\item{...}{Options to be passed to X13. See the documentation for the seas
function (https://cran.r-project.org/web/packages/seasonal/seasonal.pdf)}
}
\description{
seasthedata will automatically seasonally adjust every column in your tibble
while respecting groups. Your data must contain exactly one column of dates.
}
\details{
seasthedata is a wrapper around the seasonal library (https://github.com/christophsax/seasonal)
}
\examples{
library(dplyr)
library(tibble)
ungrouped_data <- tibble(dates = seq.Date(from = as.Date("1949-01-01"),
                         by = "month", length.out = 144),
                         y = as.vector(AirPassengers))
seasthedata(ungrouped_data)

grouped_data <- bind_rows(mutate(ungrouped_data, group = "A"),
                          mutate(ungrouped_data, group = "B"))
grouped_data <- group_by(grouped_data, group)
seasthedata(grouped_data)

}
